components/file-controller/requirements.txt
PyGithub
pygit2